Decompose script files and add setings form
