+++ DirectoryStructureToMarkdownLinks/App.config
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <configSections>
+        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+            <section name="DirectoryStructureToMarkdownLinks.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+        </sectionGroup>
+    </configSections>
+    <userSettings>
+        <DirectoryStructureToMarkdownLinks.Properties.Settings>
+            <setting name="AutoCheckMaxDepth" serializeAs="String">
+                <value>3</value>
+            </setting>
+            <setting name="DotDirExpandAllowed" serializeAs="String">
+                <value>False</value>
+            </setting>
+            <setting name="ReadmeNotificationEnabled" serializeAs="String">
+                <value>True</value>
+            </setting>
+            <setting name="AutoCheckInOpen" serializeAs="String">
+                <value>True</value>
+            </setting>
+        </DirectoryStructureToMarkdownLinks.Properties.Settings>
+    </userSettings>
+</configuration>
