+++ DirectoryStructureToMarkdownLinks/Program.cs
@@ -11,7 +11,7 @@ namespace DirectoryStructureToMarkdownLinks
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new AppForn());
+            Application.Run(new AppForm());
         }
     }
 }
